feat(lint): add tslint and fix linting errors
